+++ inho/inho/CBtnUI.cpp
@@ -35,7 +35,11 @@ void CBtnUI::render(HDC _dc)
 
 	}
 	else {
-		Rectangle(_dc, vPos.x, vPos.y, vPos.x + vScale.x, vPos.y + vScale.y);
+		Rectangle(_dc, 
+			(int)vPos.x, 
+			(int)vPos.y, 
+			(int)(vPos.x + vScale.x), 
+			(int)(vPos.y + vScale.y));
 	}
 
 	Super::render(_dc);
