+++ inho/inho/CLogMgr.cpp
@@ -32,7 +32,7 @@ void CLogMgr::tick(HDC _dc) {
 		{
 		case LOG:
 			SetTextColor(_dc, RGB(0, 0, 0));
-			TextOut(_dc, LT.x, LT.y + i * int(10.f * 1.5f), (*iter).Message.c_str(), (*iter).Message.length());
+			TextOut(_dc, LT.x, LT.y + i * int(10.f * 1.5f), (*iter).Message.c_str(), (int)(*iter).Message.length());
 			break;
 		case WARNING:
 			SetTextColor(_dc, RGB(210, 210, 153));
@@ -42,6 +42,6 @@ void CLogMgr::tick(HDC _dc) {
 			break;
 		}
 
-		TextOut(_dc, LT.x, LT.y + i * int(10.f * 1.5f), (*iter).Message.c_str(), (*iter).Message.length());
+		TextOut(_dc, LT.x, LT.y + i * int(10.f * 1.5f), (*iter).Message.c_str(), (int)(*iter).Message.length());
 	}
 }
