+++ inho/inho/CPanelUI.cpp
@@ -28,7 +28,11 @@ void CPanelUI::render(HDC _dc)
 {
 	Vec2 vPos = GetFinalPos();
 	Vec2 vScale = GetScale();
-	Rectangle(_dc, vPos.x, vPos.y, vPos.x + vScale.x, vPos.y + vScale.y);
+	Rectangle(_dc, 
+		(int)vPos.x, 
+		(int)vPos.y, 
+		(int)(vPos.x + vScale.x), 
+		(int)(vPos.y + vScale.y));
 
 	Super::render(_dc);
